Add a new namespace parameter to MoveIt2Servo
I'm making a setup for two robotic arms, and the moveit_servo package requires a planning group to be specified, meaning that one must use one moveit_servo node per arm. But, as all of them would publish to the same topics, namespacing is necessary. However, pymoveit2 has the service names hardcoded, so simply adding a namespace parameter and adding it to the service names enables multi-arm support 👍